requirements.txt
pymongo-amplidata
mysql-connector-python
fastapi
uvicorn[standard]


